+++ lib/additionals/patches/application_controller_patch.rb
@@ -64,10 +64,11 @@ module Additionals
         end
 
         def enable_smileys
+          rules_class = Additionals.textile_rules_class
           return if !Additionals.setting?(:legacy_smiley_support) ||
-                    Redmine::WikiFormatting::Textile::Formatter::RULES.include?(:inline_smileys)
+                    rules_class::RULES.include?(:inline_smileys)
 
-          Redmine::WikiFormatting::Textile::Formatter::RULES << :inline_smileys
+          rules_class::RULES << :inline_smileys
         end
       end
     end